requirements.txt
cryptography
numpy
pandas
pycoingecko
python-dotenv
requests
websockets
influxdb-client
datetime
